scenario-mcp/pyproject.toml
[project]
name = "scenario-mcp"
version = "1.0.0"
description = "Scenario.gg AI Asset Generation via Model Context Protocol"
authors = [{name = "Qusai Saleem", email = "hi@qusai.org"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.3.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "pillow>=10.0.0",
    "asyncio-throttle>=1.0.0",
    "tenacity>=8.0.0",
    "typing-extensions>=4.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"